requirements.txt
# Google Sheets Prospect Enrichment - Production Requirements

# Google API packages
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
gspread==6.0.0

# Web scraping and HTTP
requests==2.31.0
beautifulsoup4==4.12.2
urllib3==1.26.18
tldextract==5.1.1

# Data processing
python-dotenv==1.0.0

# Rich CLI interface for beautiful output
rich==13.7.0

# HTML parsing
lxml==4.9.3
html5lib==1.1

# Optional: Enhanced browser automation (only if needed)
playwright==1.41.0

# Development/Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.23.2